+++ R/gene.methylation.R
@@ -1,7 +1,8 @@
 #' Gene-level methylation
 #'
-#' Calculate gene-level methylation for a dataset containing CpGs from the Illumina 450k methylation array.
+#' Calculate gene-level methylation for a dataset containing CpGs from the Illumina 450k and 850k human methylation arrays.
 #' Gene-level methylation is calculated as the median beta-value among CpG islands in the gene promoter region.
+#' Note that gene names that originally contained a hyphen (-) are replaced with a period (.)
 #'
 #' @param methy methylation dataset where rownames give patient ids and columns use CpG ids
 #' @param print.progress TRUE/FALSE to show progress bar